+++ packages/react-rtc/src/contexts/RtcProvider.tsx
@@ -1,15 +1,9 @@
-import React, { useState, useRef, useEffect } from 'react';
+import React, { useState } from 'react';
 import Message from '../models/Message';
-import Peer from '../models/Peer';
 import { RtcContext } from './RtcContext';
-import {
-  ConnectionState,
-  type Metadata,
-  type User,
-  type Signal,
-} from '../types';
+import { type Metadata, type User } from '../types';
 import { usePubSub } from '../hooks/usePubSub';
-import { useSignaling } from '../hooks/useSignaling';
+import { usePeerConnection } from '../hooks/usePeerConnection';
 
 interface Props {
   children: JSX.Element;
@@ -22,21 +16,26 @@ export const RtcProvider = ({
   signalingServer,
   iceServers,
 }: Props) => {
-  const localUuid = useRef(crypto.randomUUID());
   const [user, setUser] = useState<User>({
     displayName: undefined,
   });
+  const handleError = (error: unknown) => dispatchEvent('error', error);
 
-  const peerConnections = useRef<Map<string, Peer>>(new Map());
   const { dispatchEvent, on, off } = usePubSub();
-  const { sendSignalingMessage, signaling } = useSignaling(localUuid.current);
+
+  const { peerConnections, onDisconnect, id, onConnect } = usePeerConnection(
+    dispatchEvent,
+    signalingServer,
+    iceServers,
+    handleError
+  );
 
   const send = (inputValue: string, metadata?: Metadata) => {
     try {
       const messageData = new Message({
         message: inputValue,
         displayName: user.displayName,
-        senderId: localUuid.current,
+        senderId: id,
         timestamp: Date.now(),
         metadata: metadata,
       });
@@ -53,185 +52,15 @@ export const RtcProvider = ({
     }
   };
 
-  function checkPeerDisconnect(peerUuid: string) {
-    const state = peerConnections.current.get(peerUuid)?.pc.iceConnectionState;
-    const peer = peerConnections.current.get(peerUuid);
-
-    if (
-      peer &&
-      (state === ConnectionState.FAILED ||
-        state === ConnectionState.CLOSED ||
-        state === ConnectionState.DISCONNECT)
-    ) {
-      dispatchEvent('peerDisconnected', peer);
-      peerConnections.current.delete(peerUuid);
-    }
-  }
-
-  function setUpPeer(peerUuid: string, displayName: string, initCall = false) {
-    const peerConnection = new RTCPeerConnection({ iceServers });
-    const dataChannel = peerConnection.createDataChannel(crypto.randomUUID());
-
-    // TODO: Pull it outside to separate file, use it as handleres
-    peerConnection.onicecandidate = (event) => gotIceCandidate(event, peerUuid);
-    peerConnection.oniceconnectionstatechange = () =>
-      checkPeerDisconnect(peerUuid);
-    peerConnection.addEventListener('datachannel', (event) =>
-      Object.defineProperty(
-        peerConnections.current.get(peerUuid),
-        'dataChannel',
-        {
-          value: event.channel,
-        }
-      )
-    );
-    peerConnection.addEventListener('connectionstatechange', () => {
-      const peer = peerConnections.current.get(peerUuid);
-      if (peer && peer.pc.connectionState === 'connected' && !initCall)
-        dispatchEvent('peerConnected', peer);
-    });
-
-    // TODO: Parse message outside, add try catch, use addMessageData
-    dataChannel.addEventListener('message', (event) =>
-      dispatchEvent('message', JSON.parse(event.data))
-    );
-
-    if (initCall) {
-      peerConnection
-        .createOffer()
-        .then((description) => createdDescription(description, peerUuid))
-        // TODO: Add error handlerer
-        .catch((e) => handleError(e));
-    }
-
-    peerConnections.current.set(
-      peerUuid,
-      new Peer({ peerConnection, dataChannel, displayName })
-    );
-  }
-
-  function gotIceCandidate(event: RTCPeerConnectionIceEvent, peerUuid: string) {
-    if (event.candidate != null) {
-      sendSignalingMessage(peerUuid, { ice: event.candidate });
-    }
-  }
-
-  function createdDescription(
-    description: RTCSessionDescriptionInit,
-
-    peerUuid: string
-  ) {
-    peerConnections.current
-      .get(peerUuid)
-      ?.pc.setLocalDescription(description)
-
-      .then(() => {
-        sendSignalingMessage(peerUuid, {
-          sdp: peerConnections.current.get(peerUuid)?.pc.localDescription,
-        });
-      })
-      .catch((e) => handleError(e));
-  }
-
-  const sendSignalingMessageToNewcomers = (uuid: string) => {
-    sendSignalingMessage(uuid, {
-      displayName: localUuid.current, // not sure if this is correct
-      uuid: localUuid.current,
-    });
-  };
-
-  const sendSessionWithDescription = (
-    peerConnection: RTCPeerConnection,
-    signal: Signal
-  ) => {
-    peerConnection
-      .setRemoteDescription(new RTCSessionDescription(signal.sdp))
-      .then(() => {
-        // Only create answers in response to offers
-        if (signal.sdp.type == 'offer') {
-          peerConnections.current
-            .get(signal.uuid)
-            ?.pc.createAnswer()
-            .then((description) => createdDescription(description, signal.uuid))
-            .catch((e) => handleError(e));
-        }
-      })
-      .catch((e) => handleError(e));
-  };
-
-  const initIceCandidate = (
-    peerConnection: RTCPeerConnection,
-    signal: Signal
-  ) => {
-    peerConnection
-      .addIceCandidate(new RTCIceCandidate(signal.ice))
-      .catch((e) => handleError(e));
-  };
-
-  function handleMessageFromServer(message: MessageEvent) {
-    const signal: Signal = JSON.parse(message.data);
-    const peerUuid = signal.uuid;
-    const peerDisplayName = signal.displayName;
-    const destination = signal.dest;
-    const isSessionDescription = signal.sdp;
-    const isIceCandidate = signal.ice;
-    // Ignore messages that are not for us or from ourselves
-    if (
-      peerUuid == localUuid.current ||
-      (destination != localUuid.current && destination != 'all')
-    )
-      return;
-
-    const currentPeerConnection = peerConnections.current.get(peerUuid)?.pc;
-
-    if (currentPeerConnection) {
-      if (isSessionDescription) {
-        sendSessionWithDescription(currentPeerConnection, signal);
-      } else if (isIceCandidate) {
-        initIceCandidate(currentPeerConnection, signal);
-      }
-    }
-
-    if (peerDisplayName) {
-      const isNewcomer = destination === localUuid.current;
-      setUpPeer(peerUuid, peerDisplayName, isNewcomer);
-      if (isNewcomer) {
-      } else {
-        sendSignalingMessageToNewcomers(peerUuid);
-      }
-    }
-  }
-
   const enter = (displayName: string, userMetadata?: Metadata) => {
-    // TODO: Fix this ignore
-
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    signaling.current = new WebSocket(signalingServer);
+    onConnect();
     setUser({ displayName, userMetadata });
   };
 
-  const disconnect = () => {
-    signaling.current?.close();
-    peerConnections.current.forEach((connection) => {
-      connection.pc.close();
-    });
-
-    // TODO: Add callback
+  const disconnect = (callback?: () => void) => {
+    onDisconnect();
+    if (callback) callback();
   };
-
-  const handleError = (error: unknown) => dispatchEvent('error', error);
-
-  useEffect(() => {
-    signaling.current?.addEventListener('message', handleMessageFromServer);
-
-    return () => {
-      signaling.current?.removeEventListener(
-        'message',
-        handleMessageFromServer
-      );
-    };
-  }, [signaling.current]);
 
   return (
     <RtcContext.Provider
